+++ cmd/gen-errors/main.go
@@ -16,6 +16,7 @@ import (
 	_ "github.com/teragrid/forge/internal/cli/cmdexplain"
 	_ "github.com/teragrid/forge/internal/cli/cmdlint"
 	_ "github.com/teragrid/forge/internal/cli/cmdnew"
+	_ "github.com/teragrid/forge/internal/cli/cmdplugin"
 	_ "github.com/teragrid/forge/internal/cli/cmdscan"
 	_ "github.com/teragrid/forge/internal/cli/cmdship"
 	_ "github.com/teragrid/forge/internal/cli/cmdupgrade"